Fix Carousel not displaying in IE
Ref: #35

The problem was caused by the `.carousel` element having the property
`position: initial;`. While this is part of the new CSS spec it is not
supported in any version of IE.

`Initial` is to set the property back to it's original value suggested
by the spec.

Changing this to `position: static;` produces the same effect as our
`div` element has a default position of `static`.